feature: splas + visual login
* Update Splash

The splash screen is complete. It includes the timer with the progress bar and a label displaying the loading percentage.
I've also added the clan and organization names; now all that's left is to give it a visually appealing look.

* Login creation + class Hash

I have created the main components for the login and I have created the class for the hash